[debug-mode]Moved code files around for better organization.
