+++ source/stellar_objects/ring.py
@@ -1,4 +1,4 @@
-from source.entity.entity import Entity
+from source.draw.entity.entity import Entity
 import math
 
 #TODO: Test this class
